Add files via upload
JeeHell FMGS is an A320 FMGS that runs outside FSX/P3D, but uses the sim's engine for the flight model and relies on the Project Airbus A320 model. There are detailed descriptions of nearly all the systems in the A320, including annunciations, buttons, switches, the MCDU, etc. It is not for the NEO, but rather the CEO. However, it may prove very useful as a reference. The Advanced User Guide contains further details, including about the flight model and FBW systems.